+++ apps/api/src/records/records.module.ts
@@ -1,9 +1,10 @@
 import { Module } from '@nestjs/common';
+import { PatientsModule } from '../patients/patients.module';
 import { RecordsController } from './records.controller';
 import { RecordsService } from './records.service';
 
 @Module({
-  imports: [],
+  imports: [PatientsModule],
   controllers: [RecordsController],
   providers: [RecordsService],
   exports: [RecordsService],